+++ include/eveio/net/UdpSocket.hpp
@@ -20,11 +20,11 @@ public:
 
   ~UdpSocket() noexcept = default;
 
-  int SendTo(StringRef data, const InetAddr &target) const noexcept;
-  int SendTo(const void *data,
+  int64_t SendTo(StringRef data, const InetAddr &target) const noexcept;
+  int64_t SendTo(const void *data,
              size_t size,
              const InetAddr &target) const noexcept;
-  int ReceiveFrom(void *buf, size_t cap, InetAddr &peer) const noexcept;
+  int64_t ReceiveFrom(void *buf, size_t cap, InetAddr &peer) const noexcept;
 };
 
 } // namespace net
