+++ include/eveio/net/UdpStream.hpp
@@ -30,11 +30,11 @@ public:
 
   ~UdpStream() noexcept;
 
-  int SendTo(StringRef data, const InetAddr &target) const noexcept;
-  int SendTo(const void *data,
+  int64_t SendTo(StringRef data, const InetAddr &target) const noexcept;
+  int64_t SendTo(const void *data,
              size_t size,
              const InetAddr &target) const noexcept;
-  int ReceiveFrom(void *buf, size_t cap, InetAddr &peer) const noexcept;
+  int64_t ReceiveFrom(void *buf, size_t cap, InetAddr &peer) const noexcept;
 
   bool SetNonblock(bool on) const noexcept;
